Fix warnings on Win build (level 2). Fix BasicCollectionProps::getStringProperty for the int and double case to actually put the int/double into the string being returned.

+++ retrieval/src/QueryEnvironment.cpp
@@ -167,8 +167,8 @@ void indri::api::QueryEnvironment::_copyStatistics( std::vector<indri::lang::Raw
     std::vector<ScoredExtentResult>& occurrencesList = statisticsResults[ scorerNodes[i]->nodeName() ][ "occurrences" ];
     std::vector<ScoredExtentResult>& contextSizeList = statisticsResults[ scorerNodes[i]->nodeName() ][ "contextSize" ];
 
-    double occurrences = UINT64(occurrencesList[0].score);
-    double contextSize = UINT64(contextSizeList[0].score);
+    double occurrences = occurrencesList[0].score;
+    double contextSize = contextSizeList[0].score;
 
     scorerNodes[i]->setStatistics( occurrences, contextSize );
   }
